requirements.txt
Pillow==8.0.1
pycryptodome==3.9.9
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2